Preserve subpixel precision: bbox/fiducial coords as Float
YOLO emits float coordinates and the affine alignment further refines
them. Casting to int at the boundary lost subpixel precision and
introduced cumulative rounding error in defect overlays.

Edge (Python)
- BoundingBox.x/y/width/height: int -> float
- DetectionItem.center_x/center_y: float division (was //)
- InspectionPacket.fiducial[12]_x/_y, DefectPayload.bbox_*: float
- yolo_detector: keep YOLO xywh as float, round to 3 decimals
- alignment._bbox_after_affine / _clip_bbox_to_image: float bounds
- crop_inspection_roi_with_offset: float math, int cast only for
  numpy slicing indices

Backend (Spring)
- InspectionRequestDto.fiducial[12]X/Y: Integer -> Float
- DefectDetailDto.bbox*: Integer -> Float (with @DecimalMin)
- InspectionResponseDto.fiducial[12]X/Y: Integer -> Float
- InspectionLog entity columns: Integer -> Float
- DefectDetail entity columns: Integer -> Float

DB schema migration handled automatically by Hibernate
ddl-auto: update (INT -> FLOAT). Existing integer values lift
cleanly into float; no data loss.

Verified end-to-end: trigger inspection -> bbox stored as
e.g. 251.387 / 982.934 (was 251 / 982).

+++ backend/src/main/java/com/inspection/dto/InspectionRequestDto.java
@@ -46,10 +46,11 @@ public class InspectionRequestDto {
     private String result;
 
     // ── 피듀셜 마크 좌표 (마크를 찾지 못한 경우 null 허용) ───────────────────
-    private Integer fiducial1X;
-    private Integer fiducial1Y;
-    private Integer fiducial2X;
-    private Integer fiducial2Y;
+    // 서브픽셀 정밀도 보존을 위해 Float
+    private Float fiducial1X;
+    private Float fiducial1Y;
+    private Float fiducial2X;
+    private Float fiducial2Y;
 
     /** 피듀셜 YOLO 탐지 신뢰도 (0~1, 선택) */
     private Float fiducial1Confidence;
